(PC-42928) feat(qpi): remove useGetCulturalSurveyContent.ts
