+++ src/features/culturalSurvey/pages/CulturalSurveyThanks.tsx
@@ -1,11 +1,24 @@
+import { useNavigation } from '@react-navigation/native'
 import React from 'react'
 
-import { useGetCulturalSurveyContent } from 'features/culturalSurvey/helpers/useGetCulturalSurveyContent'
+import { UseNavigationType } from 'features/navigation/navigators/RootNavigator/types'
 import QpiThanks from 'ui/animations/qpi_thanks.json'
 import { GenericInfoPage } from 'ui/pages/GenericInfoPage'
 
 export const CulturalSurveyThanks: React.FC = () => {
-  const { thanks } = useGetCulturalSurveyContent()
+  const { reset } = useNavigation<UseNavigationType>()
+
+  const navigateToIdentityCheckHonor = () => {
+    reset({
+      index: 0,
+      routes: [
+        {
+          name: 'SubscriptionStackNavigator',
+          state: { index: 0, routes: [{ name: 'IdentityCheckHonor' }] },
+        },
+      ],
+    })
+  }
 
   return (
     <GenericInfoPage
@@ -14,10 +27,10 @@ export const CulturalSurveyThanks: React.FC = () => {
       animationTargetShapeNames={['Fond 1', 'Gradient Fill 1']}
       animationTargetLayerNames={['étoile', 'cadre']}
       title="Un grand merci pour tes réponses&nbsp;!"
-      subtitle={thanks.subtitle}
+      subtitle="Elles nous permettent de suivre l’évolution de tes pratiques culturelles sur l’application."
       buttonPrimary={{
-        wording: thanks.button.wording,
-        onPress: thanks.button.onPress,
+        wording: 'Continuer',
+        onPress: navigateToIdentityCheckHonor,
       }}
     />
   )
